allow build time opt-out of systemd dependencies
Allow building without systemd specific code and its dependencies by
setting the 'no_systemd' flag.
