+++ cmd/cgctl/main_nosystemd.go
@@ -1,0 +1,39 @@
+// +build no_systemd
+
+/*
+   Copyright The containerd Authors.
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+package main
+
+import (
+	"github.com/urfave/cli"
+)
+
+var newSystemdCommand = cli.Command{
+	Name:  "systemd",
+	Usage: "(disabled)",
+	Action: func(clix *cli.Context) error {
+		return nil
+	},
+}
+
+var deleteSystemdCommand = cli.Command{
+	Name:  "del-systemd",
+	Usage: "(disabled)",
+	Action: func(clix *cli.Context) error {
+		return nil
+	},
+}
