+++ systemd.go
@@ -1,3 +1,5 @@
+// +build linux,!no_systemd
+
 /*
    Copyright The containerd Authors.
